requirements.txt
pandas>=2.2.0
cx_Oracle>=8.3.0
openpyxl>=3.1.2
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.18.0
PyYAML>=6.0.1
flask>=3.0.0
streamlit>=1.30.0

